+++ packages/core/src/AppSwitcher/index.js
@@ -1,0 +1,2 @@
+export { default } from "./AppSwitcher";
+export { default as HvAppSwitcherAction } from "./Action";