+++ packages/core/src/FilterGroup/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from "./FilterGroup";
+export * from "./FilterGroup";