+++ packages/core/src/index.d.ts
@@ -7,6 +7,9 @@ export * from "./ActionBar";
 export { default as HvActionsGeneric } from "./ActionsGeneric";
 export * from "./ActionsGeneric";
 
+export { default as HvAppSwitcher } from "./AppSwitcher";
+export * from "./AppSwitcher";
+
 export { default as HvAssetInventory } from "./AssetInventory";
 export * from "./AssetInventory";
 
@@ -61,6 +64,7 @@ export { default as HvFileUploader } from "./FileUploader";
 export * from "./FileUploader";
 
 export { default as HvFilterGroup } from "./FilterGroup";
+export * from "./FilterGroup";
 
 export { default as HvFooter } from "./Footer";
 export * from "./Footer";