+++ packages/lab/src/DotPagination/styles.js
@@ -1,0 +1,17 @@
+const styles = () => ({
+  root: {
+    display: "flex",
+    justifyContent: "center",
+  },
+
+  horizontal: {
+    marginLeft: 0,
+    width: "auto",
+  },
+
+  radioRoot: {
+    marginLeft: "0",
+  },
+});
+
+export default styles;