+++ packages/lab/src/index.js
@@ -16,3 +16,4 @@ export { default as HvColorPicker } from "./ColorPicker";
 export { default as HvQueryBuilder } from "./QueryBuilder";
 export { default as HvInlineEditor } from "./InlineEditor";
 export { default as HvProgressBar } from "./ProgressBar";
+export { default as HvDotPagination } from "./DotPagination";